requirements.txt
streamlit
requests
xmltodict
google-generativeai
